requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework
Pillow
requests
python-dateutil
pyimgur
Faker
django-environ
locust
celery
redis==3.4.1
